Decomposition of sources, common improvements
