+++ .gitignore
@@ -1,2 +1,3 @@
 builds
 *.pro.*
+doc
