+++ abstractgetproc.cpp
@@ -1,0 +1,6 @@
+#include "abstractgetproc.h"
+
+AbstractGETProc::AbstractGETProc()
+{
+
+}
