+++ abstractpostproc.cpp
@@ -1,0 +1,6 @@
+#include "abstractpostproc.h"
+
+AbstractPOSTProc::AbstractPOSTProc()
+{
+
+}
