+++ usergetproc.cpp
@@ -1,0 +1,6 @@
+#include "usergetproc.h"
+
+UserGETProc::UserGETProc()
+{
+
+}
