+++ userpostproc.cpp
@@ -1,0 +1,6 @@
+#include "userpostproc.h"
+
+UserPOSTProc::UserPOSTProc()
+{
+
+}
